requirements.txt
sqlparse==0.4.2
tzdata==2022.1
django-cors-headers==3.10.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.0.0
PyJWT==1.7.1
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
pytz==2021.3